requirements.txt
langchain
langchain-openai
flask
sys
python-dotenv
openai
